add pledge(2) to further lock down oftpd!
pledge(2) needs 'stdio' for writing to read()ing and write()ing to sockets, 'inet' for accept()ing connections, and 'rpath' to list directories and open() files.

I didn't bother to put error-handling if-conditions around pledge(2) because there are a finite amount of circumstances in which pledge(2) can fail and most of them are not applicable. The one circumstance that *might* cause pledge(2) to fail is if OpenBSD has at any point or will ever stop supporting the 'stdio', 'inet', or 'rpath' promises, and I don't think that ever has or will be the case.